Refactored the code, added a linter and linted the files
I moved some of the logic out of the index.js file into modules to reduce the
clutter in the index file and section the code off in a more understandable way.
I moved the refactored code under modules into a validation file which holds all
the logic to validate the password and checkPassword content which holds the
logic to check for the individual characters of the password.

More tests were added to cover the logic that was moved out of the index file.